Refactor provider test factories so it can be used in other pkgs
