+++ internal/provider/data_source_auth0_tenant_test.go
@@ -31,7 +31,7 @@ func TestAccDataTenant(t *testing.T) {
 	httpRecorder := recorder.New(t)
 
 	resource.Test(t, resource.TestCase{
-		ProviderFactories: ProviderTestFactories(httpRecorder),
+		ProviderFactories: TestFactories(httpRecorder),
 		Steps: []resource.TestStep{
 			{
 				Config: template.ParseTestName(testAccDataTenantConfig, t.Name()),
