+++ internal/provider/resource_auth0_action_test.go
@@ -109,7 +109,7 @@ func TestAccAction(t *testing.T) {
 	httpRecorder := recorder.New(t)
 
 	resource.Test(t, resource.TestCase{
-		ProviderFactories: ProviderTestFactories(httpRecorder),
+		ProviderFactories: TestFactories(httpRecorder),
 		Steps: []resource.TestStep{
 			{
 				Config: template.ParseTestName(testAccActionConfigCreateWithOnlyRequiredFields, t.Name()),
@@ -219,7 +219,7 @@ func TestAccAction_FailedBuild(t *testing.T) {
 	httpRecorder := recorder.New(t)
 
 	resource.Test(t, resource.TestCase{
-		ProviderFactories: ProviderTestFactories(httpRecorder),
+		ProviderFactories: TestFactories(httpRecorder),
 		Steps: []resource.TestStep{
 			{
 				Config: template.ParseTestName(testAccActionConfigCreateWithFailedBuild, t.Name()),
