+++ internal/provider/resource_auth0_branding_theme_test.go
@@ -187,7 +187,7 @@ func TestAccBrandingTheme(t *testing.T) {
 	httpRecorder := recorder.New(t)
 
 	resource.Test(t, resource.TestCase{
-		ProviderFactories: ProviderTestFactories(httpRecorder),
+		ProviderFactories: TestFactories(httpRecorder),
 		Steps: []resource.TestStep{
 			{
 				Config: testAccBrandingThemeCreate,
