+++ internal/provider/resource_auth0_client_grant_test.go
@@ -73,7 +73,7 @@ func TestAccClientGrant(t *testing.T) {
 	httpRecorder := recorder.New(t)
 
 	resource.Test(t, resource.TestCase{
-		ProviderFactories: ProviderTestFactories(httpRecorder),
+		ProviderFactories: TestFactories(httpRecorder),
 		Steps: []resource.TestStep{
 			{
 				Config: template.ParseTestName(testAccClientGrantConfigCreate, t.Name()),
