+++ internal/provider/resource_auth0_connection_client_test.go
@@ -68,7 +68,7 @@ func TestAccConnectionClient(t *testing.T) {
 	httpRecorder := recorder.New(t)
 
 	resource.Test(t, resource.TestCase{
-		ProviderFactories: ProviderTestFactories(httpRecorder),
+		ProviderFactories: TestFactories(httpRecorder),
 		Steps: []resource.TestStep{
 			{
 				Config: template.ParseTestName(testAccCreateConnectionClient, t.Name()),
