+++ internal/provider/resource_auth0_custom_domain_verification_test.go
@@ -17,7 +17,7 @@ func TestAccCustomDomainVerificationWithAuth0ManagedCerts(t *testing.T) {
 	httpRecorder := recorder.New(t)
 
 	resource.Test(t, resource.TestCase{
-		ProviderFactories: ProviderTestFactories(httpRecorder),
+		ProviderFactories: TestFactories(httpRecorder),
 		Steps: []resource.TestStep{
 			{
 				Config: testAccCustomDomainVerificationWithAuth0ManagedCerts,
@@ -72,7 +72,7 @@ func TestAccCustomDomainVerificationWithSelfManagedCerts(t *testing.T) {
 	httpRecorder := recorder.New(t)
 
 	resource.Test(t, resource.TestCase{
-		ProviderFactories: ProviderTestFactories(httpRecorder),
+		ProviderFactories: TestFactories(httpRecorder),
 		Steps: []resource.TestStep{
 			{
 				Config: testAccCustomDomainVerificationWithSelfManagedCerts,
