+++ internal/provider/resource_auth0_email_template_test.go
@@ -17,7 +17,7 @@ func TestAccEmailTemplate(t *testing.T) {
 	httpRecorder := recorder.New(t)
 
 	resource.Test(t, resource.TestCase{
-		ProviderFactories: ProviderTestFactories(httpRecorder),
+		ProviderFactories: TestFactories(httpRecorder),
 		Steps: []resource.TestStep{
 			{
 				Config: testAccEmailTemplateConfig,
