+++ internal/provider/resource_auth0_email_test.go
@@ -165,7 +165,7 @@ func TestAccEmail(t *testing.T) {
 	httpRecorder := recorder.New(t)
 
 	resource.Test(t, resource.TestCase{
-		ProviderFactories: ProviderTestFactories(httpRecorder),
+		ProviderFactories: TestFactories(httpRecorder),
 		Steps: []resource.TestStep{
 			{
 				Config: testAccCreateSESEmailProvider,
