+++ internal/provider/resource_auth0_organization_test.go
@@ -113,7 +113,7 @@ func TestAccOrganization(t *testing.T) {
 	httpRecorder := recorder.New(t)
 
 	resource.Test(t, resource.TestCase{
-		ProviderFactories: ProviderTestFactories(httpRecorder),
+		ProviderFactories: TestFactories(httpRecorder),
 		Steps: []resource.TestStep{
 			{
 				Config: template.ParseTestName(testAccOrganizationEmpty, strings.ToLower(t.Name())),
