+++ internal/provider/resource_auth0_prompt_custom_text_test.go
@@ -12,7 +12,7 @@ func TestAccPromptCustomText(t *testing.T) {
 	httpRecorder := recorder.New(t)
 
 	resource.Test(t, resource.TestCase{
-		ProviderFactories: ProviderTestFactories(httpRecorder),
+		ProviderFactories: TestFactories(httpRecorder),
 		Steps: []resource.TestStep{
 			{
 				Config: testAccPromptCustomTextCreate,
