+++ internal/provider/resource_auth0_prompt_test.go
@@ -42,7 +42,7 @@ func TestAccPrompt(t *testing.T) {
 	httpRecorder := recorder.New(t)
 
 	resource.Test(t, resource.TestCase{
-		ProviderFactories: ProviderTestFactories(httpRecorder),
+		ProviderFactories: TestFactories(httpRecorder),
 		Steps: []resource.TestStep{
 			{
 				Config: testAccPromptEmpty,
