+++ internal/provider/resource_auth0_rule_config_test.go
@@ -19,7 +19,7 @@ func TestAccRuleConfig(t *testing.T) {
 	httpRecorder := recorder.New(t)
 
 	resource.Test(t, resource.TestCase{
-		ProviderFactories: ProviderTestFactories(httpRecorder),
+		ProviderFactories: TestFactories(httpRecorder),
 		Steps: []resource.TestStep{
 			{
 				Config: template.ParseTestName(testAccRuleConfigCreate, t.Name()),
