+++ internal/provider/resource_auth0_rule_test.go
@@ -15,7 +15,7 @@ func TestAccRule(t *testing.T) {
 	httpRecorder := recorder.New(t)
 
 	resource.Test(t, resource.TestCase{
-		ProviderFactories: ProviderTestFactories(httpRecorder),
+		ProviderFactories: TestFactories(httpRecorder),
 		Steps: []resource.TestStep{
 			{
 				Config: template.ParseTestName(testAccRuleCreate, t.Name()),
