+++ internal/provider/resource_auth0_trigger_binding_test.go
@@ -14,7 +14,7 @@ func TestAccTriggerBinding(t *testing.T) {
 	httpRecorder := recorder.New(t)
 
 	resource.Test(t, resource.TestCase{
-		ProviderFactories: ProviderTestFactories(httpRecorder),
+		ProviderFactories: TestFactories(httpRecorder),
 		Steps: []resource.TestStep{
 			{
 				Config: template.ParseTestName(testAccTriggerBindingConfigCreate, t.Name()),
